Fixed types

+++ laplace.c
@@ -57,8 +57,8 @@ Compute the dot product of two vectors
 @param n: the size of the vectors
 @return the dot product of the two vectors
 */
-EXTERN_C double ddot(double *u, double *v, int n) {
-    int i;
+EXTERN_C double ddot(double *u, double *v, long int n) {
+    long int i;
     double result = 0.0;
     #pragma omp parallel for reduction(+:result)
     for (i = 0; i < n; i++) {
@@ -76,8 +76,8 @@ and store the result in a third vector
 @param alpha: the scaling constant
 @param n: the size of the vectors
 */
-EXTERN_C void daxpy(double *v, double *u, double *result, double alpha, int n) {
-    int i;
+EXTERN_C void daxpy(double *v, double *u, double *result, double alpha, long int n) {
+    long int i;
     #pragma omp parallel for
     for (i = 0; i < n; i++) {
         result[i] = u[i] + alpha * v[i];
@@ -92,8 +92,8 @@ and store the result in the second vector
 @param alpha: the scaling constant
 @param n: the size of the vectors
 */
-EXTERN_C void daxpy2(double *v, double *u, double alpha, int n) {
-    int i;
+EXTERN_C void daxpy2(double *v, double *u, double alpha, long int n) {
+    long int i;
     #pragma omp parallel for
     for (i = 0; i < n; i++) {
         u[i] += alpha * v[i];
@@ -106,7 +106,7 @@ Compute the Euclidean norm of a vector
 @param n: the size of the vector
 @return the Euclidean norm of the vector
 */
-EXTERN_C double norm(double *u, int n) {
+EXTERN_C double norm(double *u, long int n) {
     return sqrt(ddot(u, u, n));
 }
 
@@ -121,7 +121,7 @@ Solve the system of linear equations Ax = b using the conjugate gradient method 
 EXTERN_C int conj_grad(double *b, double *x0, double *x, double tol, int n) {
     long int i;
     long int n3 = n * n * n;
-    int iter = 0;
+    long int iter = 0;
 
     // printf("Running conjugate gradient with %d elements\n", n3);
